# docs/api/API-SPEC.yaml
openapi: 3.0.3
info:
  title: CloudBoot Core API
  description: |
    CloudBoot NG 核心API规范
    包含两部分：
    1. Boot API (Agent ↔ Core) - BootOS Agent与Server通信
    2. External API (Terraform/UI) - 外部系统集成接口
  version: 1.0.0
  contact:
    name: CloudBoot Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: 本地开发环境
  - url: http://{core_host}:{core_port}
    description: 部署环境
    variables:
      core_host:
        default: 192.168.1.100
      core_port:
        default: '8080'
...
paths:
...
  /api/v1/profiles:
    get:
      tags: [Profile]
      summary: 查询OS安装模板
      operationId: listProfiles
      responses:
        '200':
          description: 模板列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/OSProfile'

    post:
      tags: [Profile]
      summary: 创建新的OS模板
      operationId: createProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OSProfile'
      responses:
        '201':
          description: 模板创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OSProfile'
...
  /api/v1/store/import:
    post:
      tags: [Store]
      summary: 导入Provider包 (.cbp)
      description: |
        导入硬件Provider驱动包（如RAID卡驱动）。

        * **验证**: watermark签名、License检查
        * **存储**: 加密存储在Core的私有库中
        * **使用**: 任务下发时自动分发
      operationId: importProvider
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Provider导入成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  version:
                    type: string
...
  /api/stream/logs/{job_id}:
    get:
      tags: [Stream]
      summary: 实时日志流 (Server-Sent Events)
      description: |
        订阅任务的实时日志。
        浏览器保持连接，Core推送日志事件。

        * **格式**: `data: <HTML fragment>\n\n`
        * **内容**: 预格式化的HTML，包含颜色和样式
        * **更新频率**: 日志到达时立即推送（最高频率受网络影响）
      operationId: streamLogs
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SSE 流建立成功，开始推送日志
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  data: <div class="text-emerald-500">[RAID] Probing controller...</div>
                  data: <div class="text-emerald-500">[RAID] Found: LSI MegaRAID 3108</div>
                  data: <div class="text-rose-500">[RAID] ERROR: Battery low</div>
components:
  schemas:
...
    OSProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: "CentOS 7 RAID10"
        distro:
          type: string
          enum: [centos7, centos8, ubuntu20, ubuntu22, ky10, openeuler22]
        config:
          type: object
          properties:
            root_password_hash:
              type: string
            timezone:
              type: string
              example: "Asia/Shanghai"
            partitions:
              type: array
              items:
                type: object
                properties:
                  mount_point:
                    type: string
                  size:
                    type: string
                    example: "100GB"
                  fstype:
                    type: string
                    example: "ext4"
            network:
              type: object
              properties:
                dhcp:
                  type: boolean
                dns:
                  type: array
                  items:
                    type: string
            packages:
              type: array
              items:
                type: string
        created_at:
          type: string
          format: date-time